[feature/APPC-1291] - 1- Change endpoints in TopUpValuesService; 2- Create "formatDecimalPlaces" and "removeDecimalPlaces" methods and change "formatNumberWithSuffix" method so that numbers with decimal places with zeroes are removed; 3- Change from Int to String in Fiat data class parameter in TopUpDefaultValueBody to handle values with decimal places

+++ app/src/main/java/com/asfoundation/wallet/topup/TopUpFragment.kt
@@ -343,13 +343,13 @@ class TopUpFragment : DaggerFragment(), TopUpFragmentView {
   override fun setupDefaultValueChips(values: List<FiatValue>) {
     val formatter = NumberFormatterUtils.create()
     default_chip1.text =
-        values[0].symbol + formatter.formatNumberWithSuffix(values[0].amount.toLong())
+        values[0].symbol + formatter.formatNumberWithSuffix(values[0].amount.toFloat())
     default_chip2.text =
-        values[1].symbol + formatter.formatNumberWithSuffix(values[1].amount.toLong())
+        values[1].symbol + formatter.formatNumberWithSuffix(values[1].amount.toFloat())
     default_chip3.text =
-        values[2].symbol + formatter.formatNumberWithSuffix(values[2].amount.toLong())
+        values[2].symbol + formatter.formatNumberWithSuffix(values[2].amount.toFloat())
     default_chip4.text =
-        values[3].symbol + formatter.formatNumberWithSuffix(values[3].amount.toLong())
+        values[3].symbol + formatter.formatNumberWithSuffix(values[3].amount.toFloat())
   }
 
   override fun unselectChips() {
